+++ CHANGES.md
@@ -20,7 +20,7 @@ The release versions are PyPi releases.
 
 #### Fixes
   * Incorrect `os.walk` behavior if top dir is a symlink ([#240](../../issues/240).
-  * `IOError` instead of `OSError` raised in several cases ([#235](../../issues/235), [#237](../../issues/237), [#238](../../issues/238), [#241](../../issues/241), [#243](../../issues/243))
+  * `IOError` instead of `OSError` raised in several cases ([#235](../../issues/235), [#237](../../issues/237), [#238](../../issues/238), [#241](../../issues/241), [#243](../../issues/243), [#246](../../issues/246))
   * Fake `rename` did not behave like `os.rename` ([#217](../../issues/217), [#224](../../issues/224), [#216](../../issues/216), [#227](../../issues/227), [#232](../../issues/232), [#233](../../issues/233), [#239](../../issues/239))
   * Symlinks have not been considered in several functions ([#215](../../issues/215), [#210](../../issues/210), [#226](../../issues/226), [#234](../../issues/234))
   * A non-existing file that has the same as the content of the parent object was seen as existent ([#219](../../issues/219)).